Moved play method in superclass and abstracted mulligan_strategy
